// src/ui/interface.ts
export interface CurrentUser {
    id?: string;
    email?: string;
    name: string;
    permissions: string[];
    isInternal: boolean | string;
    EmployeeCode?: string;
    EmployeeData?: EmployeeData;
    isActive: boolean | string;
    createdAt: Date | string;
    updatedAt: Date | string;
}


export interface EmployeeData {
    EmployeeCode: string;
    TH_Name: string;
    ENG_Name: string;
    ADUser: string;
    Email: string;
    [key: string]: any;
}